+++ openfisca_aotearoa/variables/social_security/20/sole_parent_support.py
@@ -20,16 +20,13 @@ class social_security__eligible_for_sole_parent_support(Variable):
     value_type = bool
     entity = Person
     definition_period = MONTH
-    label = "Eligible for Job Seeker Support"
+    label = "Eligible for Sole Parent Support"
     reference = "https://www.workandincome.govt.nz/map/income-support/main-benefits/sole-parent-support/qualifications.html"
 
     def formula(persons, period, parameters):
         # The applicant
         in_nz = persons('normally_lives_in_nz', period)
-        resident_or_citizen = \
-            persons('is_resident', period) + \
-            persons('is_permanent_resident', period) + \
-            persons('is_nz_citizen', period)
+        resident_or_citizen = persons('is_citizen_or_resident', period)
 
         years_in_nz = persons('sole_parent_support__meets_years_in_nz_requirement', period)
         age_requirement = persons('sole_parent_support__meets_age_threshold', period)