+++ openfisca_aotearoa/variables/social_security/61/accommodation_supplement.py
@@ -8,15 +8,14 @@ class social_security__eligible_for_accommodation_supplement(Variable):
     definition_period = MONTH
     label = "Eligible for Accommodation Supplement"
 
-    """
-        age => 16
-        normally_lives_in_nz: True
-        is_resident: True
-        has_accomodation_costs: True
-        has_social_housing: False
-        income_below_threshold_for_accommodation_supplement: True
-        cash_below_threshold_for_accommodation_supplement: True
-    """
+    reference = """
+        61DH Purpose of accommodation supplement
+
+        The purpose of sections 61E to 61EC and Schedule 18 is to provide targeted
+        financial assistance to help certain people with high accommodation costs
+        to meet those costs.
+        """
+
     def formula(persons, period, parameters):
         # Based on MSD's web page
         # https://www.workandincome.govt.nz/products/a-z-benefits/accommodation-supplement.html
@@ -27,20 +26,13 @@ class social_security__eligible_for_accommodation_supplement(Variable):
         in_nz = persons('normally_lives_in_nz', period)
 
         resident_or_citizen = persons('is_resident', period) + persons('is_permanent_resident', period) + persons('is_nz_citizen', period)
-        has_accomodation_costs = persons('has_accomodation_costs', period)
+        social_security__has_accomodation_costs = persons('social_security__has_accomodation_costs', period)
         not_social_housing = (persons('eligible_for_social_housing', period) == 0)
 
         income = persons('accommodation_supplement__below_income_threshold', period)
         cash = persons('accommodation_supplement__below_cash_threshold', period)
 
-        return age_requirement * resident_or_citizen * in_nz * has_accomodation_costs * not_social_housing * income * cash
-
-
-class has_accomodation_costs(Variable):
-    value_type = bool
-    entity = Person
-    label = u"Has accommodation costs"
-    definition_period = MONTH
+        return age_requirement * resident_or_citizen * in_nz * social_security__has_accomodation_costs * not_social_housing * income * cash
 
 
 class eligible_for_social_housing(Variable):